Refactor checkout page to pull cart data using Java servlets
Remapped CartServlet to handle both adding to cart and getting all items to cart.

Response contains the Cart object with list of CartItem objects as a data attribute. CartItem has a Product object as one of its attributes for displaying needed information to checkout page. For video card products, the "gpu" field is needed for the display name. Since Product doesn't have a gpu attribute, CartItem also has a ProductVC attribute for video card records specifically. So Cart contains a list of CartItems that either have its Product attribute set or its ProductVC set. There should be a better way to do this, such as adding a column to the product table in the database to store displaynames.