Store customer ID server-side instead of client-side
Initially customer ID was stored client-side using sessionStorage. The ID was retrieved from sessionStorage whenever it is needed to make requests. This may be a security issue since the user has accsss to this session storage.

With this commit, the customer ID is stored server-side using HttpSession. Whenever a request requiring customerID is made, the server will retrieve the HttpSession (or create one if needed) and use the stored "customerID" attribute. The servlet code has been refactored so that "customerId" is set the same time a customer record is inserted into the database. If "customerID" has not been set when a request requiring it is made, that should mean the customer record has not been created yet. In that case, the servlet will include the CustomerServlet to create the customer record first before proceeding with the rest of the request.

Since the customerID is stored server-side, this should improve security. The Javascript code for the main/home page, product details page, and checkout page have been refactored to accommodate this change, but more refactoring is needed for the form-related code since that also uses customerID. Future commits will address this issue.